--- src/main/java/roomescape/auth/LoginRequest.java
@@ -1,8 +1,0 @@
-package roomescape.auth;
-
-public record LoginRequest (
-	String email,
-	String password
-){
-
-}
