--- src/main/java/roomescape/auth/TokenResponse.java
@@ -1,6 +1,0 @@
-package roomescape.auth;
-
-public record TokenResponse (
-	String token
-) {
-}
