+++ src/main/java/roomescape/auth/dto/request/LoginRequest.java
@@ -1,0 +1,8 @@
+package roomescape.auth.dto.request;
+
+public record LoginRequest (
+	String email,
+	String password
+){
+
+}
