+++ src/main/java/roomescape/auth/dto/response/LoginMember.java
@@ -1,0 +1,10 @@
+package roomescape.auth.dto.response;
+
+
+public record LoginMember(
+	Long id,
+	String name,
+	String email,
+	String role
+) {
+}
