+++ src/main/java/roomescape/auth/dto/response/MemberTokenDto.java
@@ -1,0 +1,8 @@
+package roomescape.auth.dto.response;
+
+public record MemberTokenDto (
+	Long id,
+	String name,
+	String role
+) {
+}
