+++ src/main/java/roomescape/auth/dto/response/TokenResponse.java
@@ -1,0 +1,6 @@
+package roomescape.auth.dto.response;
+
+public record TokenResponse (
+	String token
+) {
+}
