--- src/main/java/roomescape/auth/jwt/MemberTokenDto.java
@@ -1,8 +1,0 @@
-package roomescape.auth.jwt;
-
-public record MemberTokenDto (
-	Long id,
-	String name,
-	String role
-) {
-}
