--- src/main/java/roomescape/auth/util/DefaultTimeProvider.java
@@ -1,13 +1,0 @@
-package roomescape.auth.util;
-
-import java.util.Date;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class DefaultTimeProvider implements TimeProvider {
-	@Override
-	public Date now() {
-		return new Date();
-	}
-}
