--- src/main/java/roomescape/auth/util/TimeProvider.java
@@ -1,7 +1,0 @@
-package roomescape.auth.util;
-
-import java.util.Date;
-
-public interface TimeProvider {
-	Date now();
-}
