+++ src/main/java/roomescape/common/config/WebConfig.java
@@ -1,0 +1,24 @@
+package roomescape.common.config;
+
+import java.util.List;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import roomescape.auth.resolver.LoginMemberArgumentResolver;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+	private final LoginMemberArgumentResolver loginMemberArgumentResolver;
+
+	public WebConfig(LoginMemberArgumentResolver loginMemberArgumentResolver) {
+		this.loginMemberArgumentResolver = loginMemberArgumentResolver;
+	}
+
+	@Override
+	public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
+		argumentResolvers.add(loginMemberArgumentResolver);
+	}
+}
