+++ src/main/java/roomescape/common/util/TimeProvider.java
@@ -1,0 +1,7 @@
+package roomescape.common.util;
+
+import java.util.Date;
+
+public interface TimeProvider {
+	Date now();
+}
