--- src/main/java/roomescape/member/LoginMember.java
@@ -1,10 +1,0 @@
-package roomescape.member;
-
-
-public record LoginMember(
-	Long id,
-	String name,
-	String email,
-	String role
-) {
-}
