+++ src/main/java/roomescape/member/controller/MemberController.java
@@ -1,0 +1,39 @@
+package roomescape.member.controller;
+
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletResponse;
+import roomescape.member.service.MemberService;
+import roomescape.member.dto.request.MemberRequest;
+import roomescape.member.dto.response.MemberResponse;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.net.URI;
+
+@RestController
+public class MemberController {
+    private MemberService memberService;
+
+    public MemberController(MemberService memberService) {
+        this.memberService = memberService;
+    }
+
+    @PostMapping("/members")
+    public ResponseEntity createMember(@RequestBody MemberRequest memberRequest) {
+        MemberResponse member = memberService.createMember(memberRequest);
+        return ResponseEntity.created(URI.create("/members/" + member.getId())).body(member);
+    }
+
+    @PostMapping("/logout")
+    public ResponseEntity logout(HttpServletResponse response) {
+        Cookie cookie = new Cookie("token", "");
+        cookie.setHttpOnly(true);
+        cookie.setPath("/");
+        cookie.setMaxAge(0);
+        response.addCookie(cookie);
+        return ResponseEntity.ok().build();
+    }
+}
