+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -1,0 +1,50 @@
+package roomescape.reservation.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.net.URI;
+import java.util.List;
+
+import roomescape.auth.dto.response.LoginMember;
+import roomescape.reservation.dto.request.ReservationRequest;
+import roomescape.reservation.dto.response.ReservationResponse;
+import roomescape.reservation.service.ReservationService;
+
+@RestController
+public class ReservationController {
+
+	private final ReservationService reservationService;
+
+	public ReservationController(ReservationService reservationService) {
+		this.reservationService = reservationService;
+	}
+
+	@GetMapping("/reservations")
+	public List<ReservationResponse> list() {
+		return reservationService.findAll();
+	}
+
+	@PostMapping("/reservations")
+	public ResponseEntity create(@RequestBody ReservationRequest reservationRequest, LoginMember loginMember) {
+		if (reservationRequest.getDate() == null
+			|| reservationRequest.getTheme() == null
+			|| reservationRequest.getTime() == null) {
+			return ResponseEntity.badRequest().build();
+		}
+
+		ReservationResponse reservationResponse = reservationService.save(reservationRequest, loginMember);
+		return ResponseEntity.created(URI.create("/reservations/" + reservationResponse.getId())).body(reservationResponse);
+	}
+
+	@DeleteMapping("/reservations/{id}")
+	public ResponseEntity delete(@PathVariable Long id) {
+		reservationService.deleteById(id);
+		return ResponseEntity.noContent().build();
+	}
+}
